Add 'Download' mode to download license text
