+++ requirements.txt
@@ -5,3 +5,4 @@ reuse==1.1.2
 PyYAML
 fosslight_util>=2.1.10
 jinja2
+pyaskalono
